move utils, expand about me
