+++ components/utils/CsvConverter/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import styles from './index.module.css';
-import CopyButton from '../common/CopyButton';
+import CopyButton from '../../common/CopyButton';
 
 enum OutputFormat {
   CSV_COMMA = 'CSV_COMMA',
