+++ pages/index.tsx
@@ -1,40 +1,33 @@
 import React from 'react';
 import styles from './index.module.css';
-import Link from 'next/link';
 
 const Home: React.FC = () => {
   return (
     <main className={styles.main}>
-      <h1>Doug Jones</h1>
-      <p>Web Developer specializing in Node.js and AWS Cloud Infrastructure</p>
+      <h2>Web Developer - Toronto - Canada</h2>
 
-      <section className={styles.about}>
+      <section>
         <h3>About Me</h3>
         <p>
-          I am a passionate web developer with extensive experience in building
-          scalable applications using Node.js and AWS Cloud services. I enjoy
-          solving complex problems and continuously learning new technologies.
+          Hello, World! I`m Doug. I enjoy scalable, maintainable code and long
+          walks on the beach.
         </p>
-      </section>
-
-      <section className={styles.projects}>
-        <h3>
-          <Link href="/utils">Utilities</Link>
-        </h3>
-      </section>
-
-      <section className={styles.contact}>
-        <h3>Contact</h3>
         <p>
-          Feel free to reach out to me on{' '}
-          <a
-            href="https://www.linkedin.com/in/doug-jones2"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            LinkedIn
-          </a>{' '}
-          or <a href="mailto:th3dougler@gmail.com">email</a>.
+          I am most confident working with React, Node.js and TypeScript. I am
+          experienced in architecting, building, mantaining and scaling
+          cloud-based infrastructure. To that end, I am comfortable with both
+          relational and non relational databases, monolithic and event driven
+          microservice architectures, REST API`s, among other things.
+        </p>
+        <p>
+          I consider myself a full stack developer, including platform
+          engineering - though I will admit that my front end work is stronger
+          when I collaborate with a designer; Unless you think this website is
+          beautiful - then I guess dont worry about it
+        </p>
+        <p>
+          If youd like to know more, please check out my contact information. I
+          am not seeking work at this time, but I am always open to networking
         </p>
       </section>
     </main>
