+++ pages/utils/index.tsx
@@ -1,8 +1,8 @@
-import CsvConverter from '@/components/CsvConverter';
-import EpochConverter from '@/components/EpochConverter';
-import ListToCsv from '@/components/ListToCsv';
+import CsvConverter from '@/components/utils/CsvConverter';
+import EpochConverter from '@/components/utils/EpochConverter';
+import ListToCsv from '@/components/utils/ListToCsv';
 import styles from './index.module.css';
-import NanoidGenerator from '@/components/NanoidGenerator';
+import NanoidGenerator from '@/components/utils/NanoidGenerator';
 const UtilitiesPage = () => {
   return (
     <main className={styles.main}>
